ref: externalized getBuyHistory to /functions/invoices
